moving CMacro to ast.c_concepts and fixing some Codacy issues
